move memspace api to experimental
