+++ examples/memspace_hmat/memspace_hmat.c
@@ -1,14 +1,14 @@
 /*
  *
- * Copyright (C) 2024 Intel Corporation
+ * Copyright (C) 2024-2025 Intel Corporation
  *
  * Under the Apache License v2.0 with LLVM Exceptions. See LICENSE.TXT.
  * SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
  *
  */
 
-#include <umf/mempolicy.h>
-#include <umf/memspace.h>
+#include <umf/experimental/mempolicy.h>
+#include <umf/experimental/memspace.h>
 
 #include <numa.h>
 #include <numaif.h>
