+++ src/mempolicy.c
@@ -1,12 +1,12 @@
 /*
  *
- * Copyright (C) 2024 Intel Corporation
+ * Copyright (C) 2024-2025 Intel Corporation
  *
  * Under the Apache License v2.0 with LLVM Exceptions. See LICENSE.TXT.
  * SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
  *
  */
-#include <umf/mempolicy.h>
+#include <umf/experimental/mempolicy.h>
 
 #include "base_alloc_global.h"
 #include "mempolicy_internal.h"
