+++ src/memspace_internal.h
@@ -1,6 +1,6 @@
 /*
  *
- * Copyright (C) 2023-2024 Intel Corporation
+ * Copyright (C) 2023-2025 Intel Corporation
  *
  * Under the Apache License v2.0 with LLVM Exceptions. See LICENSE.TXT.
  * SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
@@ -10,7 +10,7 @@
 #ifndef UMF_MEMSPACE_INTERNAL_H
 #define UMF_MEMSPACE_INTERNAL_H 1
 
-#include <umf/memspace.h>
+#include <umf/experimental/memspace.h>
 
 #include "base_alloc.h"
 #include "memtarget_internal.h"
