+++ src/memspaces/memspace_lowest_latency.c
@@ -12,7 +12,7 @@
 #include <stdlib.h>
 
 #include <umf.h>
-#include <umf/memspace.h>
+#include <umf/experimental/memspace.h>
 
 // UMF_MEMSPACE_LOWEST_LATENCY requires HWLOC
 // Additionally, it is currently unsupported on Win
