+++ src/memspaces/memspace_numa.c
@@ -10,7 +10,7 @@
 #include <stdlib.h>
 
 #include <umf.h>
-#include <umf/memspace.h>
+#include <umf/experimental/memspace.h>
 
 // umfMemspaceCreateFromNumaArray requires HWLOC
 // Additionally, it is currently unsupported on Win
