+++ test/memspaces/memspace_fixtures.hpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2024 Intel Corporation
+// Copyright (C) 2024-2025 Intel Corporation
 // Under the Apache License v2.0 with LLVM Exceptions. See LICENSE.TXT.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
@@ -10,7 +10,7 @@
 #include <numaif.h>
 #include <thread>
 
-#include <umf/memspace.h>
+#include <umf/experimental/memspace.h>
 
 #include "base.hpp"
 #include "memspace_helpers.hpp"
