+++ test/memspaces/memspace_highest_capacity.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2024 Intel Corporation
+// Copyright (C) 2024-2025 Intel Corporation
 // Under the Apache License v2.0 with LLVM Exceptions. See LICENSE.TXT.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
@@ -11,7 +11,7 @@
 
 #include <numa.h>
 #include <numaif.h>
-#include <umf/memspace.h>
+#include <umf/experimental/memspace.h>
 #include <unordered_set>
 
 using umf_test::test;
