+++ test/memspaces/memspace_host_all.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2024 Intel Corporation
+// Copyright (C) 2024-2025 Intel Corporation
 // Under the Apache License v2.0 with LLVM Exceptions. See LICENSE.TXT.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
@@ -7,7 +7,7 @@
 #include <sys/mman.h>
 #include <unordered_set>
 
-#include <umf/memspace.h>
+#include <umf/experimental/memspace.h>
 
 #include "memspace_fixtures.hpp"
 #include "memspace_helpers.hpp"
